+++ docker/src/main/java/net/oneandone/stool/docker/ImageInfo.java
@@ -23,7 +23,9 @@ public class ImageInfo {
     private static final String IMAGE_PREFIX = "net.oneandone.stool-";
 
     public static final String IMAGE_LABEL_PORT_PREFIX = IMAGE_PREFIX + "port.";
-    public static final String IMAGE_LABEL_P12 = IMAGE_PREFIX + "certificate.p12";  // where to mount the generated p12 certificate into the container
+    public static final String IMAGE_LABEL_CERTIFICATE_KEY = IMAGE_PREFIX + "certificate.key";
+    public static final String IMAGE_LABEL_CERTIFICATE_CHAIN = IMAGE_PREFIX + "certificate.chain";
+    public static final String IMAGE_LABEL_CERTIFICATE_P12 = IMAGE_PREFIX + "certificate.p12";
     public static final String IMAGE_LABEL_DISK = IMAGE_PREFIX + "disk";
     public static final String IMAGE_LABEL_MEMORY = IMAGE_PREFIX + "memory";
     public static final String IMAGE_LABEL_URL_CONTEXT = IMAGE_PREFIX + "url.context";